Move tests into teradata package
