+++ plugin/trino-teradata/src/test/java/io/trino/plugin/teradata/TestLogonMechanism.java
@@ -11,9 +11,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.unit;
+package io.trino.plugin.teradata;
 
-import io.trino.plugin.teradata.LogonMechanism;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
