+++ plugin/trino-teradata/src/test/java/io/trino/plugin/teradata/integration/DatabaseConfigFactory.java
@@ -11,9 +11,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.integration;
+package io.trino.plugin.teradata.integration;
 
-import io.trino.plugin.integration.util.TeradataTestConstants;
 import io.trino.plugin.teradata.LogonMechanism;
 
 import java.util.HashMap;
