+++ plugin/trino-teradata/src/test/java/io/trino/plugin/teradata/integration/clearscape/ClearScapeManager.java
@@ -11,10 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.integration.clearscape;
+package io.trino.plugin.teradata.integration.clearscape;
 
 import io.airlift.log.Logger;
-import io.trino.plugin.integration.util.TeradataTestConstants;
+import io.trino.plugin.teradata.integration.TeradataTestConstants;
 
 import java.net.URISyntaxException;
 import java.util.regex.Pattern;
