+++ plugin/trino-teradata/src/test/java/io/trino/plugin/teradata/integration/clearscape/Error5xxException.java
@@ -11,8 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.integration.clearscape;
+package io.trino.plugin.teradata.integration.clearscape;
 
-public record GetEnvironmentRequest(
-        String name
-) {}
+public class Error5xxException
+        extends BaseException
+{
+    public Error5xxException(int statusCode, String body)
+    {
+        super(statusCode, body);
+    }
+}
