+++ plugin/trino-teradata/src/test/java/io/trino/plugin/teradata/integration/clearscape/GetEnvironmentRequest.java
@@ -11,11 +11,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.integration.util;
+package io.trino.plugin.teradata.integration.clearscape;
 
-public interface TeradataTestConstants
-{
-    String ENV_CLEARSCAPE_URL = "https://api.clearscape.teradata.com";
-    String ENV_CLEARSCAPE_USERNAME = "demo_user";
-    String ENV_CLEARSCAPE_REGION = "asia-south";
-}
+public record GetEnvironmentRequest(
+        String name
+) {}
