+++ plugin/trino-teradata/src/test/java/io/trino/plugin/teradata/integration/clearscape/Model.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.integration.clearscape;
+package io.trino.plugin.teradata.integration.clearscape;
 
 public class Model
 {
